requirements.txt

numpy>=1.24
pandas>=2.0
scikit-learn>=1.4
matplotlib>=3.8
tensorflow>=2.15
keras-tuner>=1.4.6
PyYAML>=6.0
joblib>=1.3
# dev (optional)
black>=24.0
ruff>=0.6
